Some changes to the Integration1DimGSLQAWCQAGS class: -added a function to check if singularity is inside the interval; -added another constructor for the class which takes a variable which is the minimum for the distance between the integration bounds and the singularity. If the distance is smaller than this variable, the QAGS method is used instead of the QAWC method; -this variable is initialized to zero (so the previous case). Only using the constructor can one change this quantity; -all the B0 integrations with finite 3-momentum k where initiallized with the above variable set to the ZERO variable of the file; The k=0 limit uses the previous case where the minimum distance is not considered, only if the singularity is inside the interval;
The B0 function in the k=0 limit is not converging when M1==M2. There is an endpoint singularity. This special case should be considered separetly. However, for now, I added the file's ZERO to M2 in order to make the masses different and the QAWC method is then used. In the other case the QAGS method was being used and the integral did not converge.