In this version the Aklevansky integral was reverted to the previous WRONG version without the correct integration region. This was done in order to make possible to compare the results coming from the integrated cross section to those calculated in the mark_30 code. All 3 methods agree if one considers the WRONG A integral, the same parameters and the same precisions.
This means that: THIS CODE GIVES THE SAME RESULTS FOR THE INTEGRATED CROSS SECTIONS AS THE PREVIOUS mark_30 IF ONE REVERTS BACK TO THE WRONG A INTEGRAL!

The correct AKlev integral is commented and the main.cpp cotains several tests. In the next version the main.cpp will be cleaned and the wrong version of the AKlev integral will be deleted for cleaness purposes. If one whishes to compare with mark_30 in the future, the Aklev must be found in previous versions.